+++ magik-checks/src/main/java/org/stevenlooman/sw/magik/checks/FileNotInLoadListCheck.java
@@ -1,7 +1,6 @@
 package org.stevenlooman.sw.magik.checks;
 
 import com.sonar.sslr.api.AstNode;
-import com.sonar.sslr.api.AstNodeType;
 
 import org.sonar.check.Rule;
 import org.stevenlooman.sw.magik.MagikCheck;
@@ -10,7 +9,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Collections;
 import java.util.List;
 import javax.annotation.Nullable;
 
@@ -19,11 +17,6 @@ public class FileNotInLoadListCheck extends MagikCheck {
 
   private static final String MESSAGE = "File is not included in load_list.";
   public static final String CHECK_KEY = "FileNotInLoadList";
-
-  @Override
-  public List<AstNodeType> subscribedTo() {
-    return Collections.emptyList();
-  }
 
   @Override
   public void visitFile(@Nullable AstNode node) {