+++ magik-checks/src/main/java/org/stevenlooman/sw/magik/checks/SyntaxErrorCheck.java
@@ -1,13 +1,8 @@
 package org.stevenlooman.sw.magik.checks;
 
 import com.sonar.sslr.api.AstNode;
-import com.sonar.sslr.api.AstNodeType;
 import org.sonar.check.Rule;
 import org.stevenlooman.sw.magik.MagikCheck;
-import org.stevenlooman.sw.magik.api.MagikGrammar;
-
-import java.util.Arrays;
-import java.util.List;
 
 @Rule(
     key = SyntaxErrorCheck.CHECK_KEY,
@@ -19,13 +14,7 @@ public class SyntaxErrorCheck extends MagikCheck {
   private static final String MESSAGE = "Invalid code.";
 
   @Override
-  public List<AstNodeType> subscribedTo() {
-    return Arrays.asList(
-        MagikGrammar.SYNTAX_ERROR);
-  }
-
-  @Override
-  public void visitNode(AstNode node) {
+  protected void walkPreSyntaxError(AstNode node) {
     addIssue(MESSAGE, node.getToken());
   }