+++ CHANGES.md
@@ -1,10 +1,11 @@
 Changes
 =======
 
-0.4.1 (unreleased)
+0.5.0 (unreleased)
 
 - MessagePatchGrammar no longer needs a $ at the end of the message patch
 - Strip org.sonarsource.sonarqube:sonar-plugin-api from magik-lint, shaving size
+- Update MagikGrammar for easier processing + introduce AstWalker
 
 
 0.4.0 (2020-01-29)